+++ dataAnalysis/helperFunctions/helper_functions_new.py
@@ -2180,7 +2180,7 @@ def catSpikesGenerator(nBins = None, type = 'RMS', timePoint = None, subSet = sl
     calcBounds = bounds is None
     if type == 'RMS':
         def catSpikes(spikes, chanIdx):
-            rms = np.sqrt(np.mean(np.array(spikes['Waveforms'][chanIdx][:,subSet], dtype = np.float32) ** 2, axis = 1))
+            rms = np.sqrt(np.mean(np.array(spikes['Waveforms'][chanIdx][:,subSet], dtype = float) ** 2, axis = 1))
             if calcBounds:
                 bounds = np.linspace(rms.min() * 0.99, rms.max() * 1.01, nBins + 1)
             return np.digitize(rms, bounds)
